fix: Integrate semantic-release with build process
The version is driven by semantic-release.
In order to use it in the package (both exported to NPM
and as header metadata in the signature), we build the
package in the `postversion` step of NPM scripts.

- The CD no longer needs its own build step
- Version is read from package.json (when updated)

However, in order for the package to export its version as
as standalone string (rather than runtime reading of wrong
environment variables or whatnot), a pre-build script is
used to generate the version tag, also containing the short
Git SHA-1 (see `generateVersion.ts`).