feat: Use semantic-release
Version is generated twice, depending on the deployment target:
- For NPM, version is generated at post-install
- For the deployed bundle, we use the latest tag

By running semantic-release for NPM first, we (may)
create a new tag to indicate the current version.
However to inject the correct version into the NPM
package, we need the Git SHA1, which is injected into
a file packed within the package. The post-install script
will then concatenate the package.json version with
the Git SHA1 and rewrite the version.js and version.d.ts
files in ./lib.

For the bundle, we now read the version not from package.json,
but from the latest tag, as semantic-release might
not always run (eg: chore commits like deps updates).
We also inject the first 8 characters of the signature
public key for easier identification.